requirements.txt
pandas
matplotlib
seaborn
pefile
yara-python
requests